modify. change id type to Long

+++ src/main/java/com/app/repository/TradeRecordRepository.java
@@ -15,7 +15,7 @@ import com.app.model.User;
 @Repository
 public interface TradeRecordRepository extends JpaRepository<TradeRecord, Integer>{
 	@Query(value="select * from TradeRecord where sellerId=%?1 or buyerId=%?1",nativeQuery=true)
-	List<TradeRecord> findByTrader(int traderId,Pageable pageable);
+	List<TradeRecord> findByTrader(Long traderId,Pageable pageable);
 	List<TradeRecord> findByTradeCard(Card Card,Pageable pageable);
 }
